backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mangum
boto3
supabase
